# backend/static/openapi.yaml
openapi: 3.0.3
info:
  title: GNN Classifier API
  version: 0.1.1
  description: |
    Default system and infrastructure endpoints.
    Includes health checks, connectivity tests, and user management.
servers:
  - url: http://127.0.0.1:8000
    description: Local development server
  - url: http://20.89.168.233:8000
    description: Hosted backend server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ClassificationResult:
      type: object
      properties:
        label:
          type: string
        confidence:
          type: number
          format: float
        subject:
          type: string
        subject_code:
          type: string
        grade:
          type: integer
        processing_time_ms:
          type: number
          format: float
        top_predictions:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              confidence:
                type: number
                format: float
              subject:
                type: string
              subject_code:
                type: string
              grade:
                type: integer
    ClassificationHistoryItem:
      type: object
      properties:
        id:
          type: integer
        image_path:
          type: string
        label:
          type: string
        confidence:
          type: number
          format: float
        subject:
          type: string
        subject_code:
          type: string
        grade:
          type: integer
        model_variant:
          type: string
        created_at:
          type: string
          format: date-time
paths:
...
  /model/classification:
    post:
      summary: Run image classification
      description: Upload an image, classify it, store the image and classification history
      tags: [Model]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
      parameters:
        - in: header
          name: X-Model-Variant
          required: false
          schema:
            type: string
            enum:
              - kNN-Voting
              - GraphSAGE-I_v2
              - GraphSAGE-E_kNN
          description: Model variant to use
      responses:
        "200":
          description: Classification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassificationResult"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
        "503":
          description: Model temporarily unavailable
  /model/classifications:
    get:
      summary: Get user classification history
      description: >
        Returns paginated classification history.
        Supports optional search by public code or subject.
      tags: [Model]
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number

        - in: query
          name: q
          required: false
          schema:
            type: string
          description: >
            Optional search keyword.
            Matches public code or subject.

      responses:
        "200":
          description: Paginated classification history
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ClassificationHistoryItem"
                  page:
                    type: integer
                    example: 1
                  limit:
                    type: integer
                    example: 3
                  total:
                    type: integer
                    example: 12
                  total_pages:
                    type: integer
                    example: 4
                  q:
                    type: string
                    nullable: true
                    example: ENG
        "401":
          description: Unauthorized
...